implement the payment process

+++ Payments.Application/Commands/CreatePayment/CreatePaymentCommand.cs
@@ -3,5 +3,5 @@ using Core.Result;
 
 namespace Payments.Application.Commands.CreatePayment
 {
-    public record CreatePaymentCommand(Guid OrderId, decimal Amount, Guid PaymentMethodId, Guid StatusId) : IRequest<Result<Guid>>;
+    public record CreatePaymentCommand(Guid OrderId, double Amount, Guid PaymentMethodId, Guid StatusId) : IRequest<Result<Guid>>;
 }
