+++ Payments.Application/Commands/UpdatePayment/UpdatePaymentCommand.cs
@@ -3,5 +3,5 @@ using Core.Result;
 
 namespace Payments.Application.Commands.UpdatePayment
 {
-    public record UpdatePaymentCommand(Guid Id, Guid OrderId, decimal Amount, Guid PaymentMethodId, Guid StatusId) : IRequest<Result<bool>>;
+    public record UpdatePaymentCommand(Guid Id, Guid OrderId, double Amount, Guid PaymentMethodId, Guid StatusId) : IRequest<Result<bool>>;
 }
